+++ spec/requests/api/customer/thanks_spec.rb
@@ -2,6 +2,7 @@ require 'rails_helper'
 include ActionController::RespondWith
 
 RSpec.describe "Api::Customer::Thanks", type: :request do
+  include LoginSupport
 
   before(:each) do
     customer = build(:customer)
@@ -11,7 +12,6 @@ RSpec.describe "Api::Customer::Thanks", type: :request do
     @office = create(:office, :with_staffs)
     @staff  = Staff.find_by(office_id: @office.id)
     @thank  = build(:thank, user: @customer, office: @office, staff: @staff)
-
     specialist = build(:specialist)
     specialist.skip_confirmation!
     specialist.save
@@ -22,9 +22,7 @@ RSpec.describe "Api::Customer::Thanks", type: :request do
   context 'ログイン済み' do
     context 'カスタマー' do
       it "お礼を作成できる" do
-	    	login(@customer)
-	    	auth_params = get_auth_params_from_login_response_headers(response)
-
+        auth_params = login(@customer)
         expect {
           post api_office_thanks_path(@thank.office_id),
           params: {
@@ -41,9 +39,7 @@ RSpec.describe "Api::Customer::Thanks", type: :request do
 
     context 'ケアマネ' do
       it "お礼を作成できない" do
-	    	login(@specialist)
-	    	auth_params = get_auth_params_from_login_response_headers(response)
-
+        auth_params = login(@customer)
         expect {
           post api_office_thanks_path(@thank.office_id),
           params: {
@@ -72,29 +68,5 @@ RSpec.describe "Api::Customer::Thanks", type: :request do
         }
       }.to change(@specialist.thanks, :count).by(0)
     end
-  end
-
-	def login(user)
-    post api_login_path,
-    params: { email: user.email, password: 'password' }
-    .to_json,
-    headers: { 'CONTENT_TYPE' => 'application/json', 'ACCEPT' => 'application/json' }
-	end
-
-  def get_auth_params_from_login_response_headers(response)
-    client = response.headers['client']
-    token = response.headers['access-token']
-    expiry = response.headers['expiry']
-    token_type = response.headers['token-type']
-    uid = response.headers['uid']
-
-    auth_params = {
-      'access-token' => token,
-      'client' => client,
-      'uid' => uid,
-      'expiry' => expiry,
-      'token-type' => token_type
-    }
-    auth_params
   end
 end
